[Decomp][Spell] Cohesion split: extract HandleAuraDummy to SpellAuraDummy.cpp
Begin the SpellAuras.cpp decomposition by moving the large catch-all
Aura::HandleAuraDummy handler (the per-spell-id script-driven dummy aura
monster, ~1400 lines) out into a new SpellAuraDummy.cpp translation unit.

Pure file-level cohesion split: the declaration stays in SpellAuras.h, the
AuraHandler[] dispatch table stays in SpellAuras.cpp and resolves the moved
member method by address at link time. Byte-identical move, no behaviour
change. SpellAuras.cpp 12994 -> 11593.
